pyproject.toml
[build-system]
requires = [ "setuptools",]
build-backend = "setuptools.build_meta"

[project]
name = "reachy-mini-agora-web-sdk"
version = "0.2.2"
description = ""
readme = "README_EN.md"
requires-python = ">=3.10"
dependencies = [ "python-dotenv", "reachy_mini_dances_library", "reachy-mini>=1.3.1", "requests>=2.32.0", "agora-python-server-sdk>=2.2.1", "fastapi>=0.115.0", "uvicorn>=0.30.0",]
[[project.authors]]
name = "Pollen Robotics"
email = "contact@pollen-robotics.com"

[dependency-groups]
dev = [ "pytest", "pytest-asyncio", "ruff==0.12.0", "mypy==1.18.2", "pre-commit", "types-requests", "python-semantic-release>=10.5.3",]

[project.scripts]
reachy-mini-agora-web-sdk = "reachy_mini_agora_web_sdk.main:main"
reachy-mini-agora-web-sdk-server = "reachy_mini_agora_web_sdk.web_rtc_server:run_web_rtc_server"

[tool.setuptools]
include-package-data = true

[tool.ruff]
line-length = 119
exclude = [ ".venv", "dist", "build", "**/__pycache__", "*.egg-info", ".mypy_cache", ".pytest_cache",]

[tool.mypy]
python_version = "3.12"
files = [ "src/",]
ignore_missing_imports = true
strict = true
show_error_codes = true
warn_unused_ignores = true

[project.entry-points.reachy_mini_apps]
reachy_mini_agora_web_sdk = "reachy_mini_agora_web_sdk.main:ReachyMiniAgoraConversationApp"

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff.lint]
select = [ "E", "F", "W", "I", "C4", "D",]
ignore = [ "E501", "D100", "D203", "D213",]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.setuptools.packages.find]
where = [ "src",]

[tool.ruff.lint.isort]
length-sort = true
lines-after-imports = 2
no-lines-before = [ "standard-library", "local-folder",]
known-local-folder = [ "reachy_mini_agora_web_sdk",]
known-first-party = [ "reachy_mini", "reachy_mini_dances_library",]
split-on-trailing-comma = true
